created SettingsModal to generalize/modularize settings section of code in Squat exercise
